+++ spec/controllers/components_controller_spec.rb
@@ -3,15 +3,17 @@ require 'rails_helper'
 RSpec.describe ComponentsController, type: :controller do
   render_views
   
-  let!(:skater) { create(:skater)}
-  let!(:competition) { create(:competition) }
+  let!(:men_skater) { create(:skater)}
+  let!(:ladies_skater) { create(:skater, :ladies)}  
+  let!(:world) { create(:competition) }
+  let!(:finlandia) { create(:competition, :finlandia) }  
   let!(:short_ss){
-    short = competition.scores.create(segment: "SHORT", skater: skater)
-    short.components.create(number: 1, name: "Skating Skill", value: 10.0)
+    short = world.scores.create(category: "MEN", segment: "SHORT", skater: men_skater, ranking: 1)
+    short.components.create(number: 1, name: "Skating Skill", factor: 1.0, value: 10.0)
   }
   let!(:free_tr){
-    free = competition.scores.create(segment: "FREE", skater: skater)
-    free.components.create(number: 2, name: "Transitions", value: 9.0)
+    free = finlandia.scores.create(category: "LADIES", segment: "FREE", skater: ladies_skater, ranking: 2)
+    free.components.create(number: 2, name: "Transitions", factor: 1.8, value: 9.0)
   }
   
   context 'index' do
@@ -19,54 +21,60 @@ RSpec.describe ComponentsController, type: :controller do
       get :index
       expect(response).to be_success
     end
-=begin
     
     it 'lists SkatingSkill' do
       get :list, xhr: true
-      expect(response.body).to include('Skating Skill')
-      expect(response.body).to include('10.0')
+      expect(response).to be_succes
+      expect_to_include(short_ss.name)
     end
-    it 'filters by segment' do
-      get :list, xhr: true, params: { segment: "FREE"}
-      expect(response.body).to include('Skating Skill')
-      expect(response.body).to include('9.0')
-      expect(response.body).not_to include('10.0')
+
+    attrs = [:competition_name, :category, :segment, :season, :ranking, :skater_name, :nation, :name, :number, :factor, :value]
+    context 'filter: ' do
+      attrs.each do |key|
+        it key do
+          get :list, xhr: true, params: { segment: "SHORT"}
+          expect_to_include(short_ss.name)
+          expect_not_to_include(free_tr.name)
+        end
+      end
     end
+
+
+    context 'sort:' do
+      attrs.each do |key|
+        it key do
+          expect_order(short_ss, free_tr, key)
+        end
+      end
+    end
+    
     context 'value comparison' do
+      it 'compares by >' do
+        get :list, xhr: true, params: { value: "9.5", value_operator: '>'}
+        expect_to_include(short_ss.name)
+        expect_not_to_include(free_tr.name)
+      end
       it 'compares by <' do
         get :list, xhr: true, params: { value: "9.5", value_operator: '<'}
-        expect(response.body).to include('Skating Skill')
-        expect(response.body).to include('9.0')
-        expect(response.body).not_to include('10.0')
+        expect_to_include(free_tr.name)
+        expect_not_to_include(short_ss.name)
       end
       it 'compares by <=' do
         get :list, xhr: true, params: { value: "9", value_operator: '<='}
-        expect(response.body).to include('Skating Skill')
-        expect(response.body).to include('9.0')
-        expect(response.body).not_to include('10.0')
+        expect_to_include(free_tr.name)
+        expect_not_to_include(short_ss.name)
+
       end
       it 'compares by >=' do
         get :list, xhr: true, params: { value: "9", value_operator: '>='}
-        expect(response.body).to include('Skating Skill')
-        expect(response.body).to include('9.0')
-        expect(response.body).to include('10.0')
+        expect_to_include(short_ss.name)
+        expect_to_include(free_tr.name)
       end
       it 'compares by =' do
         get :list, xhr: true, params: { value: "9", value_operator: '='}
-        expect(response.body).to include('Skating Skill')
-        expect(response.body).to include('9.0')
-        expect(response.body).not_to include('10.0')
-      end
-    end
-=end
-  end
-=begin
-  context 'sort:' do
-    [:competition_name, :category, :segment, :season, :ranking, :skater_name, :nation, :name, :number, :factor, :value].each do |key|
-      it key do
-        expect_order(short_ss, free_tr, key)
+        expect_to_include(free_tr.name)
+        expect_not_to_include(short_ss.name)
       end
     end
   end
-=end
 end