+++ spec/controllers/skaters_controller_spec.rb
@@ -1,12 +1,6 @@
 require 'rails_helper'
 
 RSpec.describe SkatersController, type: :controller do
-  def expect_to_include_skater(skater)
-    [:name, :category, :isu_number].each do |key|
-      expect_to_include(skater[key])
-    end
-  end
-  
   render_views
 
   let!(:men_skater){ create(:skater) }
@@ -31,26 +25,21 @@ RSpec.describe SkatersController, type: :controller do
 
     it 'list' do
       get :list, xhr: true
-      expect_to_include_skater(men_skater)
-      expect_to_include_skater(ladies_skater)
-      expect_not_to_include('No SCORES')
+      expect_to_include(men_skater.name)
+      expect_to_include(ladies_skater.name)
+      expect_not_to_include(no_scores_skater.name)
         
     end
+    attrs = [:name, :category, :nation]
     context 'filters: ' do
-      [:name, :category, :nation].each do |key|
-        it do
-          get :list, xhr: true, params: { key => men_skater.send(key) }
-          expect_to_include_skater(men_skater)
-          expect_not_to_include(no_scores_skater.name)
-
-          get :list, xhr: true, params: filter_params(key, men_skater.send(key))
-          expect_to_include_skater(men_skater)
-          expect_not_to_include(no_scores_skater.name)
+      attrs.each do |key|
+        it key do
+          expect_filter(men_skater, ladies_skater, key)
         end
       end
     end
     context 'sort: ' do
-      [:name, :category, :nation].each do |key|
+      attrs.each do |key|
         it key do
           expect_order(men_skater, ladies_skater, key)
         end
@@ -60,8 +49,8 @@ RSpec.describe SkatersController, type: :controller do
       [:json, :csv].each do |format|
         it format do
           get :index, params: { format: format }
-          expect_to_include_skater(men_skater)
-          expect_to_include_skater(ladies_skater)
+          expect_to_include(men_skater.name)
+          expect_to_include(ladies_skater.name)
         end
       end
     end
@@ -70,16 +59,16 @@ RSpec.describe SkatersController, type: :controller do
   context 'show: ' do
     it 'isu_number' do
       get :show, params: { isu_number: men_skater.isu_number }
-      expect_to_include_skater(men_skater)
+      expect_to_include(men_skater.name)
     end
     it 'name' do
       get :show, params: { isu_number: men_skater.name }
-      expect_to_include_skater(men_skater)
+      expect_to_include(men_skater.name)
     end
     it 'json' do
-      get :show, params: { isu_number: 1, format: 'json' }
+      get :show, params: { isu_number: men_skater.isu_number, format: :json }
       expect(response.content_type).to eq("application/json")
-      expect_to_include_skater(men_skater)
+      expect_to_include(men_skater.name)
     end
   end
 end